Working but no populating
Okay if this is stable then ive got the envelope creation and sending working. Most of it seems to be configured within the docusign template and the routes set in my code, though i should do a better job at understanding whats going on in there. I can create and send a docusign to my own email but i still havent gotten the autopopulation working. That is what we are trying to fix today.